CRUD done

+++ mongos/exapp1/students.controller.js
@@ -15,6 +15,7 @@ const setupRoutes = (app) => {
     res.send(student);
   });
 
+  // TODO: Proper searching with paging and ordering
   app.post("/api/students/search", async (req, res) => {
     console.log("POST /api/students", req.body);
     const result = await search(req.body);
@@ -39,16 +40,31 @@ const setupRoutes = (app) => {
 
   app.put("/api/students/update/:id", async (req, res) => {
     console.log("PUT /api/students/:id", req.params.id);
-    const updated = await update(req.params.id, req.body);
-    res.send(updated);
-    // console.log("req", req.body);
-    // res.send("thank you");
+    const validationResult = validate(req.body);
+    if (req.params.id && !validationResult.error) {
+      const result = await update(req.params.id, req.body);
+      if (result instanceof Error) {
+        res.status(400).json(JSON.parse(result.message));
+        return;
+      }
+      return res.json(result);
+    }
+    return res
+      .status(400)
+      .json({ status: "error", message: validationResult.error });
   });
 
   app.delete("/api/students/delete/:id", async (req, res) => {
     console.log("DELETE /api/students/:id", req.params.id);
-    const result = await deleteById(req.params.id);
-    res.send(result);
+    if (req.params.id) {
+      const result = await deleteById(req.params.id);
+      if (result instanceof Error) {
+        res.status(400).json(JSON.parse(result.message));
+        return;
+      }
+      return res.json(result);
+    }
+    return res.status(400).json({ status: "error", message: "Id required" });
   });
 };
